% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData1.R
\docType{data}
\name{simData1}
\alias{simData1}
\title{Simulated univariate functional data for method FADPclust}
\format{fd, see FDA R package for details.}
\description{
Simulated univariate functional data, with 2 clusters each containing 100 sample curves, were for users to apply the method FADPclust.
}
\keyword{datasets}
