% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData2.R
\docType{data}
\name{simData2}
\alias{simData2}
\title{Simulated multivariate functional data for method FADPclust}
\format{fd, see FDA R package for details.}
\description{
Simulated three-dimensional multivariate functional data, with 2 clusters each containing 100 sample curves, were for users to apply the method FADPclust.
}
\keyword{datasets}
